requirements.txt
asgiref==3.2.10
astroid==2.4.2
confusable-homoglyphs==3.2.0
Django==3.0.7
django-bootstrap3==14.0.0
django-registration==3.1
django-tinymce==3.0.2
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
